requirements.txt
django==5.2.4
djangorestframework==3.15.2
psycopg2===2.9.10
